requirements.txt
aes==1.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==22.6.0
fastapi==0.110.0
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
httptools==0.6.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
joblib==1.3.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pika==1.3.2
pluggy==1.4.0
protobuf==4.25.2
pyasn1==0.5.1
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
starlette==0.36.3
threadpoolctl==3.3.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0